rename indexing module
