+++ src/zarr/codecs/sharding.py
@@ -34,8 +34,7 @@ from zarr.common import (
     parse_shapelike,
     product,
 )
-from zarr.indexing import c_order_iter, morton_order_iter
-from zarr.indexing2 import BasicIndexer
+from zarr.indexing import BasicIndexer, c_order_iter, morton_order_iter
 from zarr.metadata import ArrayMetadata, parse_codecs
 
 if TYPE_CHECKING:
